bimtester: steps, split steps testmodule
