+++ src/ifcbimtester/bimtester/features/steps/attributes_eleclasses.py
@@ -18,22 +18,6 @@ def step_impl(context, ifc_class, pattern):
             assert False
 
 
-@step("all {ifc_class} elements have a {qto_name}.{quantity_name} quantity")
-def step_impl(context, ifc_class, qto_name, quantity_name):
-    elements = IfcFile.get().by_type(ifc_class)
-    for element in elements:
-        is_successful = False
-        if not element.IsDefinedBy:
-            assert False
-        for relationship in element.IsDefinedBy:
-            if relationship.RelatingPropertyDefinition.Name == qto_name:
-                for quantity in relationship.RelatingPropertyDefinition.Quantities:
-                    if quantity.Name == quantity_name:
-                        is_successful = True
-        if not is_successful:
-            assert False
-
-
 @step('there is an {ifc_class} element with a {attribute_name} attribute with a value of "{attribute_value}"')
 def step_impl(context, ifc_class, attribute_name, attribute_value):
     elements = IfcFile.get().by_type(ifc_class)
@@ -55,33 +39,6 @@ def step_impl(context, ifc_class, attribute):
     for element in elements:
         if not getattr(element, attribute):
             assert False
-
-
-@step("all (?P<ifc_class>.*) elements have an? (?P<property_path>.*\..*) property")
-def step_impl(context, ifc_class, property_path):
-    pset_name, property_name = property_path.split(".")
-    elements = IfcFile.get().by_type(ifc_class)
-    for element in elements:
-        if not IfcFile.get_property(element, pset_name, property_name):
-            assert False
-
-
-@step(
-    'all (?P<ifc_class>.*) elements have an? (?P<property_path>.*\..*) property value matching the pattern "(?P<pattern>.*)"'
-)
-def step_impl(context, ifc_class, property_path, pattern):
-    import re
-
-    pset_name, property_name = property_path.split(".")
-    elements = IfcFile.get().by_type(ifc_class)
-    for element in elements:
-        prop = IfcFile.get_property(element, pset_name, property_name)
-        if not prop:
-            assert False
-        # For now, we only check single values
-        if prop.is_a("IfcPropertySingleValue"):
-            if not (prop.NominalValue and re.search(pattern, prop.NominalValue.wrappedValue)):
-                assert False
 
 
 @step('all (?P<ifc_class>.*) elements have an? (?P<attribute>.*) matching the pattern "(?P<pattern>.*)"')
